move schema error types
